updated some stuff to run on CCR
