+++ .gitignore
@@ -1,9 +1,14 @@
 # Do not track these directories
 data/*.json
-saved_models/
-best_models/
 archived/
 figures/
+wandb/
+saved_models/CGCNN/
+saved_models/Painn/
+saved_models/e3nn/
+best_models/CGCNN/
+best_models/Painn/
+best_models/e3nn/
 
 
 # Byte-compiled / optimized / DLL files
