--- docs/MIGRATION_GUIDE_SIGOPT_TO_WANDB.md
@@ -1,1 +1,0 @@
-# moved file
