+++ inference/test_model_prediction.py
@@ -7,7 +7,8 @@ import time
 from processing.dataloader.dataloader import get_dataloader
 from processing.utils import filter_data_by_properties,select_structures
 from processing.interpolation.Interpolation import *
-from training.sigopt_utils import build_sigopt_name
+# from training.sigopt_utils import build_sigopt_name  # Original SigOpt utils (commented out)
+from training.wandb_utils import build_wandb_name  # Wandb utils (active)
 from processing.create_model.create_model import create_model
 from inference.select_best_models import get_experiment_id
 from nff.train.loss import build_mae_loss
@@ -100,11 +101,20 @@ def get_model_prediction(test_set_type, model_params, gpu_num, target_prop, num_
     test_loader = get_dataloader(test_data,target_prop,model_type,1,interpolation,per_site=per_site,long_range=model_params["long_range"])       
     end_2 = time.time()
     
-    sigopt_name = build_sigopt_name(model_params["data"], target_prop, model_params["struct_type"], model_params["interpolation"], model_params["model_type"],contrastive_weight=model_params["contrastive_weight"],training_fraction=model_params["training_fraction"],long_range=model_params["long_range"])
+    # Original SigOpt name building (commented out)
+    # sigopt_name = build_sigopt_name(model_params["data"], target_prop, model_params["struct_type"], model_params["interpolation"], model_params["model_type"],contrastive_weight=model_params["contrastive_weight"],training_fraction=model_params["training_fraction"],long_range=model_params["long_range"])
+    # exp_id = get_experiment_id(model_params, target_prop)
+    
+    # Wandb name building (active)
+    wandb_name = build_wandb_name(model_params["data"], target_prop, model_params["struct_type"], model_params["interpolation"], model_params["model_type"],contrastive_weight=model_params["contrastive_weight"],training_fraction=model_params["training_fraction"],long_range=model_params["long_range"])
     exp_id = get_experiment_id(model_params, target_prop)
 
     for idx in range(num_best_models):
-        directory = "./best_models/" + model_params["model_type"] + "/" + sigopt_name + "/" +str(exp_id) + "/" + "best_" + str(idx)
+        # Original SigOpt directory (commented out)
+        # directory = "./best_models/" + model_params["model_type"] + "/" + sigopt_name + "/" +str(exp_id) + "/" + "best_" + str(idx)
+        
+        # Wandb directory (active)
+        directory = "./best_models/" + model_params["model_type"] + "/" + wandb_name + "/" +str(exp_id) + "/" + "best_" + str(idx)
         model, normalizer = load_model(gpu_num, train_loader, model_params, directory, target_prop,per_site=per_site)
         start_3 = time.time()
         prediction = evaluate_model_with_tracked_ids(model, normalizer, gpu_num, test_loader, model_params)
@@ -198,12 +208,9 @@ def evaluate_model_with_tracked_ids(model, normalizer, gpu_num, test_loader, mod
 
                 predictions_iter = normalizer.denorm(output).detach().cpu().numpy().reshape(crys_idx.shape[0],-1)
                 
-
-                for i in range(crys_idx.shape[0]):
-                    predictions[int(crys_idx[i])] = predictions_iter[i]
-                    all_crys_ids.append(int(crys_idx[i]))
+                for i in range(len(crys_idx)):
+                    predictions[crys_idx[i]] = predictions_iter[i]
 
         if return_ids:
-            return predictions, all_crys_ids
-
+            return predictions, list(predictions.keys())
         return predictions
