+++ training/run_wandb_experiment.py
@@ -9,16 +9,12 @@ import random
 import shutil
 import os
 import wandb
-# Original SigOpt imports (commented out)
-# import sigopt
 from processing.utils import filter_data_by_properties,select_structures
 from processing.interpolation.Interpolation import *
 from processing.dataloader.dataloader import get_dataloader
 from processing.create_model.create_model import create_model
 from training.hyperparameters.wandb_parameters import *
 from training.model_training.trainer import *
-# Original SigOpt utils (commented out)
-# from training.sigopt_utils import build_sigopt_name
 from training.wandb_utils import build_wandb_name
 from training.evaluate import *
 
@@ -26,10 +22,6 @@ from training.evaluate import *
 def run_wandb_experiment(struct_type,model_type,gpu_num,experiment_id=None,parallel_band=1,obs_budget=50,training_fraction=1.0,data_name="data/",target_prop="dft_e_hull",interpolation=False,contrastive_weight=1.0,training_seed=0,nickname=""):
     """Run wandb hyperparameter optimization experiment"""
     
-    # Original SigOpt code (commented out)
-    # if experiment_id == None:
-    #     sigopt_settings={"parallel_band": parallel_band,"obs_budget": obs_budget}
-
     if data_name == "data/":
         training_data = pd.read_json(data_name + 'training_set.json')
         training_data = training_data.sample(frac=training_fraction,replace=False,random_state=training_seed)
@@ -57,61 +49,6 @@ def run_wandb_experiment(struct_type,model_type,gpu_num,experiment_id=None,paral
 
     print("Completed data processing")
 
-    # Original SigOpt code (commented out)
-    # conn = sigopt.Connection(driver="lite")
-    # sigopt_name = build_sigopt_name(data_name,target_prop,struct_type,interpolation,model_type,contrastive_weight,training_fraction,training_seed)
-    # 
-    # if experiment_id == None:
-    #     experiment = create_sigopt_experiment(data_name,target_prop,struct_type,interpolation,model_type,contrastive_weight,training_fraction,training_seed,sigopt_settings,conn)
-    #     print("Created a new SigOpt experiment '" + sigopt_name + "' with ID: " + str(experiment.id))
-    # else:
-    #     experiment = conn.experiments(experiment_id).fetch()
-    #     print("Continuing a prior SigOpt experiment '" + sigopt_name + "' with ID: " + str(experiment.id))
-    # 
-    # while experiment.progress.observation_count < experiment.observation_budget:
-    #     print('\n========================\nSigopt experiment count #', experiment.progress.observation_count)
-    #     suggestion = conn.experiments(experiment.id).suggestions().create()
-    #     hyperparameters = suggestion.assignments
-    # 
-    #     value = sigopt_evaluate_model(data_name,suggestion.assignments,processed_data,target_prop,interpolation,struct_type,model_type,contrastive_weight,training_fraction,training_seed,experiment.id,experiment.progress.observation_count,gpu_num,nickname)    
-    #     training_results = {"validation_loss": value}
-    # 
-    #     conn.experiments(experiment.id).observations().create(
-    #         suggestion=suggestion.id,
-    #         values=[{"name": "val_mae", "value": value}],
-    #     )
-    # 
-    #     experiment = conn.experiments(experiment.id).fetch()
-    #     observation_id = experiment.progress.observation_count - 1
-    # 
-    #     model_save_dir = './saved_models/'+ model_type + '/' + sigopt_name + '/' + str(experiment.id) + '/observ_' + str(observation_id)
-    #     model_tmp_dir = './saved_models/'+ model_type + '/' + sigopt_name + '/' + str(experiment_id) + '/' + nickname + '_tmp' + str(gpu_num)
-    # 
-    #     with open(model_tmp_dir + '/hyperparameters.json', 'w') as file:
-    #         json.dump(hyperparameters, file)
-    #     with open(model_tmp_dir + '/training_results.json', 'w') as file:
-    #         json.dump(training_results, file)
-    # 
-    #     if not os.path.exists(model_save_dir):
-    #         os.makedirs(model_save_dir)
-    # 
-    #     ### Copy contents of tmp file
-    #     possible_file_names = ["best_model", "best_model.pth.tar", "best_model.torch",
-    #                            "final_model.torch","final_model","final_model.pth.tar",
-    #                            "log_human_read.csv","checkpoints/checkpoint-100.pth.tar",
-    #                            'hyperparameters.json','training_results.json']
-    #     for file_name in possible_file_names:
-    #         if os.path.isfile(model_tmp_dir + "/" + file_name):
-    #             if file_name == "checkpoints/checkpoint-100.pth.tar":
-    #                 shutil.move(model_tmp_dir + "/" + file_name, model_save_dir + "/" + "checkpoint-100.pth.tar")
-    #             else:
-    #                 shutil.move(model_tmp_dir + "/" + file_name, model_save_dir + "/" + file_name)
-    #     
-    #     ### Empty tmp file
-    #     shutil.rmtree(model_tmp_dir)
-    #     torch.cuda.empty_cache()
-
-    # Wandb equivalent (active)
     # Create wandb sweep configuration
     sweep_config = create_wandb_experiment(data_name,target_prop,struct_type,interpolation,model_type,contrastive_weight,training_fraction,training_seed,obs_budget)
     wandb_name = build_wandb_name(data_name,target_prop,struct_type,interpolation,model_type,contrastive_weight,training_fraction,training_seed)
@@ -194,36 +131,6 @@ def run_wandb_experiment(struct_type,model_type,gpu_num,experiment_id=None,paral
 
 def wandb_evaluate_model(data_name,hyperparameters,processed_data,target_prop,interpolation,struct_type,model_type,contrastive_weight,training_fraction,training_seed,experiment_id,observation_count,gpu_num,nickname):
     """Evaluate model for wandb experiment"""
-    
-    # Original SigOpt evaluation code (commented out)
-    # device = "cuda:" + str(gpu_num)
-    # train_data = processed_data[0]
-    # validation_data = processed_data[1]
-    # per_site = False
-    # if "per_site" in target_prop:
-    #     per_site = True
-    # train_loader = get_dataloader(train_data,target_prop,model_type,hyperparameters["batch_size"],interpolation,per_site=per_site)
-    # train_eval_loader = None
-    # if "e3nn" in model_type and "pretrain" not in data_name and "per_site" not in target_prop:
-    #     train_eval_loader = get_dataloader(train_data,target_prop,"e3nn_contrastive",1,interpolation,per_site=per_site)
-    #     val_loader = get_dataloader(validation_data,target_prop,"e3nn_contrastive",1,interpolation,per_site=per_site)
-    # else:
-    #     val_loader = get_dataloader(validation_data,target_prop,model_type,1,interpolation,per_site=per_site)
-    # model, normalizer = create_model(model_type,train_loader,interpolation,target_prop,hyperparameters=hyperparameters,per_site=per_site)
-    # sigopt_name = build_sigopt_name(data_name,target_prop,struct_type,interpolation,model_type,contrastive_weight,training_fraction,training_seed)
-    # model_tmp_dir = './saved_models/'+ model_type + '/' + sigopt_name + '/' + str(experiment_id) + '/' + nickname + '_tmp' + str(gpu_num)
-    # if os.path.exists(model_tmp_dir):
-    #     shutil.rmtree(model_tmp_dir)
-    # os.makedirs(model_tmp_dir) 
-    # best_model,loss_fn = trainer(model,normalizer,model_type,train_loader,val_loader,hyperparameters,model_tmp_dir,gpu_num,train_eval_loader=train_eval_loader,contrastive_weight=contrastive_weight)
-    # is_contrastive = False
-    # if "contrastive" in model_type:
-    #     is_contrastive = True
-    # _, _, best_loss = evaluate_model(best_model, normalizer, model_type, val_loader, loss_fn, gpu_num,is_contrastive=is_contrastive, contrastive_weight=contrastive_weight)
-    # if model_type == "Painn":
-    #     return best_loss
-    # else:
-    #     return best_loss[0]
 
     # Wandb equivalent (active)
     device = "cuda:" + str(gpu_num)
@@ -283,23 +190,6 @@ def wandb_evaluate_model(data_name,hyperparameters,processed_data,target_prop,in
 
 def create_wandb_experiment(data_name,target_prop,struct_type,interpolation,model_type,contrastive_weight,training_fraction,training_seed,obs_budget):
     """Create wandb sweep configuration"""
-    
-    # Original SigOpt experiment creation (commented out)
-    # sigopt_name = build_sigopt_name(data_name,target_prop,struct_type,interpolation,model_type,contrastive_weight,training_fraction,training_seed)
-    # if model_type == "Painn":
-    #     curr_parameters = get_painn_hyperparameter_range()
-    # elif model_type == "CGCNN":
-    #     curr_parameters = get_cgcnn_hyperparameter_range()
-    # else:
-    #     curr_parameters = get_e3nn_hyperparameter_range()
-    # experiment = conn.experiments().create(
-    #     name=sigopt_name, 
-    #     parameters = curr_parameters,
-    #     metrics=[dict(name="val_mae", objective="minimize", strategy="optimize")],
-    #     observation_budget=sigopt_settings["obs_budget"], 
-    #     parallel_bandwidth=sigopt_settings["parallel_band"],
-    # )
-    # return experiment
 
     # Wandb equivalent (active)
     wandb_name = build_wandb_name(data_name,target_prop,struct_type,interpolation,model_type,contrastive_weight,training_fraction,training_seed)
